+++ lib/ui/viewer/gallery/component/group/lazy_group_gallery.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:math';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -10,11 +9,11 @@ import 'package:photos/models/file.dart';
 import 'package:photos/models/selected_files.dart';
 import 'package:photos/theme/ente_theme.dart';
 import "package:photos/ui/viewer/gallery/component/day_widget.dart";
-import 'package:photos/ui/viewer/gallery/component/grid/lazy_loading_grid_view.dart';
 import 'package:photos/ui/viewer/gallery/component/grid/place_holder_grid_view_widget.dart';
+import "package:photos/ui/viewer/gallery/component/group/group_gallery.dart";
 import 'package:photos/ui/viewer/gallery/gallery.dart';
 
-class LazyLoadingGallery extends StatefulWidget {
+class LazyGroupGallery extends StatefulWidget {
   final List<File> files;
   final int index;
   final Stream<FilesUpdatedEvent>? reloadEvent;
@@ -28,7 +27,7 @@ class LazyLoadingGallery extends StatefulWidget {
   final int photoGridSize;
   final bool enableFileGrouping;
   final bool limitSelectionToOne;
-  LazyLoadingGallery(
+  LazyGroupGallery(
     this.files,
     this.index,
     this.reloadEvent,
@@ -46,10 +45,10 @@ class LazyLoadingGallery extends StatefulWidget {
   }) : super(key: key ?? UniqueKey());
 
   @override
-  State<LazyLoadingGallery> createState() => _LazyLoadingGalleryState();
+  State<LazyGroupGallery> createState() => _LazyGroupGalleryState();
 }
 
-class _LazyLoadingGalleryState extends State<LazyLoadingGallery> {
+class _LazyGroupGalleryState extends State<LazyGroupGallery> {
   static const kNumberOfDaysToRenderBeforeAndAfter = 8;
 
   late Logger _logger;
@@ -168,7 +167,7 @@ class _LazyLoadingGalleryState extends State<LazyLoadingGallery> {
   }
 
   @override
-  void didUpdateWidget(LazyLoadingGallery oldWidget) {
+  void didUpdateWidget(LazyGroupGallery oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (!listEquals(_files, widget.files)) {
       _reloadEventSubscription?.cancel();
@@ -232,7 +231,7 @@ class _LazyLoadingGalleryState extends State<LazyLoadingGallery> {
           ],
         ),
         _shouldRender!
-            ? GetGallery(
+            ? GroupGallery(
                 photoGridSize: widget.photoGridSize,
                 files: _files,
                 tag: widget.tag,
@@ -242,6 +241,8 @@ class _LazyLoadingGalleryState extends State<LazyLoadingGallery> {
                 areAllFromDaySelected: _areAllFromDaySelected,
                 limitSelectionToOne: widget.limitSelectionToOne,
               )
+            // todo: perf eval should we have separate PlaceHolder for Groups
+            //  instead of creating a large cached view
             : PlaceHolderGridViewWidget(
                 _files.length,
                 widget.photoGridSize,
@@ -256,58 +257,5 @@ class _LazyLoadingGalleryState extends State<LazyLoadingGallery> {
     } else {
       _showSelectAllButton.value = true;
     }
-  }
-}
-
-class GetGallery extends StatelessWidget {
-  final int photoGridSize;
-  final List<File> files;
-  final String tag;
-  final GalleryLoader asyncLoader;
-  final SelectedFiles? selectedFiles;
-  final ValueNotifier<bool> toggleSelectAllFromDay;
-  final ValueNotifier<bool> areAllFromDaySelected;
-  final bool limitSelectionToOne;
-  const GetGallery({
-    required this.photoGridSize,
-    required this.files,
-    required this.tag,
-    required this.asyncLoader,
-    required this.selectedFiles,
-    required this.toggleSelectAllFromDay,
-    required this.areAllFromDaySelected,
-    required this.limitSelectionToOne,
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    const kRecycleLimit = 400;
-    final List<Widget> childGalleries = [];
-    final subGalleryItemLimit = photoGridSize * subGalleryMultiplier;
-
-    for (int index = 0; index < files.length; index += subGalleryItemLimit) {
-      childGalleries.add(
-        LazyLoadingGridView(
-          tag,
-          files.sublist(
-            index,
-            min(index + subGalleryItemLimit, files.length),
-          ),
-          asyncLoader,
-          selectedFiles,
-          index == 0,
-          files.length > kRecycleLimit,
-          toggleSelectAllFromDay,
-          areAllFromDaySelected,
-          photoGridSize,
-          limitSelectionToOne: limitSelectionToOne,
-        ),
-      );
-    }
-
-    return Column(
-      children: childGalleries,
-    );
   }
 }
